requirements.txt
asgiref==3.8.1
attrs==25.3.0
bleach==6.2.0
Django==5.2.3
djangorestframework==3.16.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
psycopg2-binary==2.9.10
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
webencodings==0.5.1
djangorestframework-simplejwt